# api/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_erhan_falcon_api_internal_models.ScopeItemJSON:
    properties:
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: integer
      kind:
        type: string
      last_run_at:
        type: string
      next_run_at:
        type: string
      program_id:
        type: integer
      schedule_cron:
        type: string
      source_url:
        type: string
      value:
        type: string
    type: object
...
  internal_handlers.errorResp:
    properties:
      error:
        type: string
    type: object
...
  internal_handlers.pageResp:
    properties:
      items:
        items: {}
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  internal_handlers.runStartResp:
    properties:
      run_id:
        type: integer
      status:
        type: string
    type: object
  internal_handlers.scopeReq:
    properties:
      enabled:
        type: boolean
      kind:
        description: domain|wildcard|ip|cidr
        type: string
      schedule_cron:
        type: string
      value:
        type: string
    type: object
...
info:
  contact: {}
  description: Bug bounty & continuous recon backend.
  title: Falcon API
  version: 0.1.0
paths:
...
  /api/programs/{id}/scope:
    get:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: scannable | all
        in: query
        name: scope
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.pageResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: List scope items
      tags:
      - scope
    post:
      consumes:
      - application/json
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      - description: scope item body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.scopeReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.ScopeItemJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Create scope item
      tags:
      - scope
  /api/programs/{id}/scope/run-all:
    post:
      parameters:
      - description: program id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Trigger pipeline for every scope item in a program
      tags:
      - scope
...
  /api/scope/{id}:
    delete:
      parameters:
      - description: scope item id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Delete scope item
      tags:
      - scope
    patch:
      consumes:
      - application/json
      parameters:
      - description: scope item id
        in: path
        name: id
        required: true
        type: integer
      - description: fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.scopeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_erhan_falcon_api_internal_models.ScopeItemJSON'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Update scope item
      tags:
      - scope
  /api/scope/{id}/run:
    post:
      parameters:
      - description: scope item id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/internal_handlers.runStartResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Trigger pipeline run
      tags:
      - scope
...
  /api/vrt:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Bugcrowd VRT taxonomy
      tags:
      - vrt
  /api/vrt/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/internal_handlers.errorResp'
      security:
      - BearerAuth: []
      summary: Refresh VRT from upstream
      tags:
      - vrt
securityDefinitions:
  BearerAuth:
    description: 'JWT token issued by /api/auth/login. Format: "Bearer {token}".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"